# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /registration:
    get:
      operationId: getRegistration
      tags: [registration]
      summary: Get the signed-in user's registration profile
      responses:
        "200":
          description: Registration status and community participation access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationStatus"
        "401":
          description: Authentication is required
    put:
      operationId: saveRegistration
      tags: [registration]
      summary: Create or update the signed-in user's registration profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "200":
          description: Saved registration profile and community participation access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationStatus"
        "400":
          description: Invalid registration details
        "401":
          description: Authentication is required
...
  /places/{cityId}/{listingSource}/corrections:
    post:
      operationId: submitListingCorrection
      tags: [listings]
      summary: Submit a guest correction for a public listing
      description: |
        Resolves the listing server-side and records a correction for review. The
        submission never changes the public listing. A required Idempotency-Key
        makes retries safe; reusing a key with different input returns 409.
      parameters:
        - name: cityId
          in: path
          required: true
          schema: { type: string, minLength: 2, maxLength: 20 }
        - name: listingSource
          in: path
          required: true
          schema:
            type: string
            enum: [google_maps, openstreetmap, curated, source_scan]
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 8, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitListingCorrectionInput"
      responses:
        "201":
          description: Correction accepted for review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingCorrectionReceipt"
        "200":
          description: Exact replay of an earlier submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingCorrectionReceipt"
        "400":
          description: Invalid input or unknown listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: Listing was not found in the server-side catalogue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: Idempotency key was reused with different input
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "429":
          description: Guest submission rate limit exceeded
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /listing-corrections/moderation:
    get:
      operationId: getListingCorrectionQueue
      tags: [moderation]
      summary: List pending guest listing corrections
      responses:
        "200":
          description: Pending corrections ordered oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingCorrectionReviewItem"
        "403":
          description: Reviewer authorization required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /listing-corrections/{correctionId}/decision:
    post:
      operationId: decideListingCorrection
      tags: [moderation]
      summary: Approve or reject a guest listing correction
      parameters:
        - name: correctionId
          in: path
          required: true
          schema: { type: number, minimum: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecideListingCorrectionInput"
      responses:
        "200":
          description: Correction decision recorded; public listing is unchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingCorrectionReviewItem"
        "403":
          description: Reviewer authorization required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: Stale version or correction is no longer pending
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
...
  schemas:
...
    BusinessLookupMatch:
      type: object
      description: Public-only business match. Never contains contact details, claimant data, or pending-claim details.
      required: [kind, cityId, listingSource, listingId, name, isClaimed]
      properties:
        kind:
          type: string
          enum: [listing, profile]
          description: "`listing` comes from the current public listing source; `profile` is a business already known to Buurtplaza."
        cityId:
          type: string
        listingSource:
          type: string
        listingId:
          type: string
        name:
          type: string
        neighborhood:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        isClaimed:
          type: boolean
          description: A verified representative already manages this business; a new claim would be recorded as disputed.
...
    ApiError:
      type: object
      description: |
        Stable, safe error shape for account and lifecycle operations. `code` is machine-readable,
        `messageKey` maps to a localized message on the client, and `correlationId` matches the
        server log entry. Never contains personal data or provider details. Existing operations
        keep their `{ error }` string responses for compatibility.
      required: [code, messageKey, correlationId]
      properties:
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        messageKey:
          type: string
          description: Localization key, e.g. `errors.auth_required`.
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/ApiFieldError"
        correlationId:
          type: string
        expectedVersion:
          type: number
          description: Present on VERSION_CONFLICT responses; the current server version the client must reload before retrying.
        duplicateCandidates:
          type: array
          description: Present on DUPLICATE_CANDIDATES responses; public matches that may already represent the submitted new business.
          items:
            $ref: "#/components/schemas/BusinessLookupMatch"
    ApiErrorCode:
      type: string
      enum:
        - AUTH_REQUIRED
        - EMAIL_UNVERIFIED
        - ACCOUNT_SUSPENDED
        - ACCOUNT_DELETED
        - FORBIDDEN
        - SELF_REVIEW_FORBIDDEN
        - NOT_FOUND
        - FEATURE_DISABLED
        - VALIDATION_FAILED
        - UNKNOWN_FIELD
        - VERSION_CONFLICT
        - IDEMPOTENCY_CONFLICT
        - DUPLICATE_CANDIDATES
        - RATE_LIMITED
        - PROVISIONING_UNAVAILABLE
        - DEPENDENCY_UNAVAILABLE
    ApiFieldError:
      type: object
      required: [field, code]
      properties:
        field:
          type: string
        code:
          type: string
          description: Machine-readable per-field reason, e.g. `required`, `unknown`, `not_in_controlled_list`.
...
    RegistrationType:
      type: string
      enum: [consumer, business]
    RegistrationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        registrationType:
          $ref: "#/components/schemas/RegistrationType"
        email:
          type: string
          pattern: "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
        usefulnessRating:
          type: number
          minimum: 1
          maximum: 5
        referralLikelihood:
          type: number
          minimum: 1
          maximum: 5
        desiredFeatures:
          type: string
          minLength: 1
          maxLength: 2000
      required: [name, registrationType, email, usefulnessRating, referralLikelihood, desiredFeatures]
    Registration:
      allOf:
        - $ref: "#/components/schemas/RegistrationInput"
        - type: object
          properties:
            createdAt:
              type: string
            updatedAt:
              type: string
          required: [createdAt, updatedAt]
    RegistrationStatus:
      type: object
      properties:
        registered:
          type: boolean
        canParticipate:
          type: boolean
        registration:
          oneOf:
            - $ref: "#/components/schemas/Registration"
            - type: "null"
      required: [registered, canParticipate, registration]
...
    SubmitListingCorrectionInput:
      type: object
      additionalProperties: false
      properties:
        listingId:
          type: string
          minLength: 1
          maxLength: 240
        fieldKey:
          type: string
          enum: [name, address, neighborhood, website_url, opening_hours, category, accessibility, dietary, price]
        proposedValue:
          type: string
          minLength: 1
          maxLength: 2000
        explanation:
          type: string
          maxLength: 4000
        evidenceUrl:
          type: string
          pattern: '^https?://.+'
          maxLength: 2048
        locale:
          type: string
          enum: [nl, en]
        consentNoticeVersion:
          type: string
          minLength: 1
          maxLength: 64
      required: [listingId, fieldKey, proposedValue, locale, consentNoticeVersion]
    ListingCorrectionReceipt:
      type: object
      properties:
        receipt:
          type: string
          description: Opaque public receipt; it is not a database identifier.
        status:
          type: string
          enum: [pending_review]
        cityId:
          type: string
        listingSource:
          type: string
        listingId:
          type: string
        fieldKey:
          type: string
        submittedAt:
          type: string
          format: date-time
      required: [receipt, status, cityId, listingSource, listingId, fieldKey, submittedAt]
    DecideListingCorrectionInput:
      type: object
      additionalProperties: false
      properties:
        decision:
          type: string
          enum: [approve, reject]
        expectedVersion:
          type: number
          minimum: 1
        reason:
          type: string
          maxLength: 2000
      required: [decision, expectedVersion]
    ListingCorrectionReviewItem:
      type: object
      properties:
        id: { type: number }
        receipt: { type: string }
        cityId: { type: string }
        listingSource: { type: string }
        listingId: { type: string }
        fieldKey: { type: string }
        proposedValue: { type: string }
        explanation: { type: string, nullable: true }
        evidenceUrl: { type: string, nullable: true }
        locale: { type: string }
        status: { type: string, enum: [pending_review, approved, rejected] }
        version: { type: number }
        reason: { type: string, nullable: true }
        reviewedBy: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        reviewedAt: { type: string, format: date-time, nullable: true }
      required: [id, receipt, cityId, listingSource, listingId, fieldKey, proposedValue, explanation, evidenceUrl, locale, status, version, reason, reviewedBy, createdAt, reviewedAt]
...