# lib/api-zod/src/generated/types/apiErrorCode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ApiErrorCode = typeof ApiErrorCode[keyof typeof ApiErrorCode];


export const ApiErrorCode = {
  AUTH_REQUIRED: 'AUTH_REQUIRED',
  EMAIL_UNVERIFIED: 'EMAIL_UNVERIFIED',
  ACCOUNT_SUSPENDED: 'ACCOUNT_SUSPENDED',
  ACCOUNT_DELETED: 'ACCOUNT_DELETED',
  FORBIDDEN: 'FORBIDDEN',
  SELF_REVIEW_FORBIDDEN: 'SELF_REVIEW_FORBIDDEN',
  NOT_FOUND: 'NOT_FOUND',
  FEATURE_DISABLED: 'FEATURE_DISABLED',
  VALIDATION_FAILED: 'VALIDATION_FAILED',
  UNKNOWN_FIELD: 'UNKNOWN_FIELD',
  VERSION_CONFLICT: 'VERSION_CONFLICT',
  IDEMPOTENCY_CONFLICT: 'IDEMPOTENCY_CONFLICT',
  DUPLICATE_CANDIDATES: 'DUPLICATE_CANDIDATES',
  RATE_LIMITED: 'RATE_LIMITED',
  PROVISIONING_UNAVAILABLE: 'PROVISIONING_UNAVAILABLE',
  DEPENDENCY_UNAVAILABLE: 'DEPENDENCY_UNAVAILABLE',
} as const;