# lib/api-zod/src/generated/types/apiFieldError.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ApiFieldError {
  field: string;
  /** Machine-readable per-field reason, e.g. `required`, `unknown`, `not_in_controlled_list`. */
  code: string;
}