# lib/api-zod/src/generated/types/registrationType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RegistrationType = typeof RegistrationType[keyof typeof RegistrationType];


export const RegistrationType = {
  consumer: 'consumer',
  business: 'business',
} as const;